hal_hw_interface:  Add hal_hw_interface.ros_hal_component module
This module provides a class representing a HAL user component running
in a ROS node, with the boilerplate taken care of.

HAL components built with this class only need to define the component
name and implement an init and update function.  When used with the
pin objects from the `hal_hw_interface.ros_hal_pin` module, useful
components can be written in just a dozen lines.