hal_hw_interface:  Add `reset` HAL pin to controller manager
When the `reset` pin goes high, call `switch_controller()` on all
active controllers.  This is intended to reset command to feedback to
avoid jumps in command, e.g. when enabling drives for the first time
after power on (where feedback will be nv encoder value, command will
be zero).

This must happen outside the thread that calls `update()`, since that
function takes care of switching the controlers, so the periodic
`reset` pin check happens from a timer in the (non-RT) spinner with
hard-coded value.  Thanks to @destogl for
this information.

This obviously can be improved; the hard-coded value needs a way to be
passed in before node init, and if the `reset` function needs to be
RT, it can be exposed in a separate HAL function.