hal_hw_interface:  Remove hal_control_node sleep() call
Somewhere along the way, this weird problem was resolved and the
controller manager doesn't segfault without the call to `sleep()`.

It's not clear if it was fixed in this project or an upstream package,
but because this shouldn't have been needed in the first place, I
won't bother with a postmortem.